Added --sort and --report flags, children usage feature
