+++ package.json
@@ -7,6 +7,7 @@
     "preversion": "npm test",
     "postversion": "git push && git push --tags && npm publish",
     "start": "node src/jsx-info.js",
+    "test-cli" : "node src/jsx-info.js --directory ./temp",
     "test": "eslint src && prettier --list-different \"src/**/*.js\""
   },
   "repository": {
